Fix view's bug.
Fix bug that won't load the table's rows after changing to different
section.

[Note]
    Some unknown behaviour is found in view's model after changing into
different section. That is, when the view is initialized for the first
time at page load, each model have a 'collection' attribute.
    But when the view is created for the second time (or more), the
'collection' attribute is gone. It probably can cause some bugs at the
future, probably. It is not fixed yet.